refactor: split internal/message/manager.go by domain
Pure move: manager.go (1126 lines, 48 declarations) split into
concern-named sibling files (manager_areas_store, manager_areas,
manager_areas_order, manager_post, manager_read, manager_index),
following the same-package sibling-file convention used by
internal/config/config_*.go and internal/menu/executor_*.go. Only
import blocks changed; no behavior, exported API, or test changes.
