+++ internal/message/manager.go
@@ -1,16 +1,13 @@
 package message
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"log/slog"
 	"os"
 	"path/filepath"
-	"sort"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/ViSiON-3/vision-3-bbs/internal/jam"
 )
@@ -139,956 +136,6 @@ func (mm *MessageManager) tearlineForNetwork(network string) string {
 // Close is a no-op now that bases are opened on-demand.
 // Kept for API compatibility.
 func (mm *MessageManager) Close() error {
-	return nil
-}
-
-// loadMessageAreas loads area definitions from JSON.
-func (mm *MessageManager) loadMessageAreas() error {
-	data, err := os.ReadFile(mm.areasPath)
-	if err != nil {
-		return err
-	}
-	if len(data) == 0 {
-		return nil
-	}
-
-	var areasList []*MessageArea
-	if err := json.Unmarshal(data, &areasList); err != nil {
-		return fmt.Errorf("failed to unmarshal areas from %s: %w", mm.areasPath, err)
-	}
-
-	mm.mu.Lock()
-	defer mm.mu.Unlock()
-
-	mm.areasByID = make(map[int]*MessageArea)
-	mm.areasByTag = make(map[string]*MessageArea)
-	mm.areasByEchoTag = make(map[string]*MessageArea)
-
-	for _, area := range areasList {
-		if area == nil {
-			continue
-		}
-		if _, exists := mm.areasByID[area.ID]; exists {
-			slog.Warn("duplicate area ID; skipping", "id", area.ID)
-			continue
-		}
-		mm.areasByID[area.ID] = area
-		mm.areasByTag[area.Tag] = area
-		// Also index by EchoTag for FTN inbound routing when tag-prefix is in use.
-		if area.EchoTag != "" && area.EchoTag != area.Tag {
-			mm.areasByEchoTag[area.EchoTag] = area
-		}
-		slog.Debug("loaded area", "id", area.ID, "tag", area.Tag, "type", area.AreaType)
-	}
-
-	// Migration: assign positions to any areas that have Position <= 0.
-	// Finds the current max position and assigns sequentially after it.
-	maxPos := 0
-	hasUnset := false
-	for _, area := range mm.areasByID {
-		if area.Position > maxPos {
-			maxPos = area.Position
-		}
-		if area.Position <= 0 {
-			hasUnset = true
-		}
-	}
-	if hasUnset && len(mm.areasByID) > 0 {
-		sorted := make([]*MessageArea, 0, len(mm.areasByID))
-		for _, area := range mm.areasByID {
-			if area.Position <= 0 {
-				sorted = append(sorted, area)
-			}
-		}
-		sort.Slice(sorted, func(i, j int) bool {
-			return sorted[i].ID < sorted[j].ID
-		})
-		for _, area := range sorted {
-			maxPos++
-			area.Position = maxPos
-		}
-		slog.Info("auto-assigned message area positions (migration)", "count", len(sorted))
-	}
-
-	return nil
-}
-
-// openBase opens a JAM base on-demand. The caller must close it when done.
-// This method does not hold any locks and should be called after releasing mm.mu.
-// Returns ErrAreaNotFound if the area doesn't exist.
-func (mm *MessageManager) openBase(areaID int) (*jam.Base, *MessageArea, error) {
-	mm.mu.RLock()
-	area, exists := mm.areasByID[areaID]
-	mm.mu.RUnlock()
-
-	if !exists {
-		return nil, nil, ErrAreaNotFound
-	}
-
-	basePath := mm.resolveBasePath(area)
-	b, err := jam.Open(basePath)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to open JAM base for area %d: %w", areaID, err)
-	}
-
-	return b, area, nil
-}
-
-// resolveBasePath returns the absolute path for a JAM base.
-func (mm *MessageManager) resolveBasePath(area *MessageArea) string {
-	bp := area.BasePath
-	if bp == "" {
-		bp = "msgbases/" + strings.ToLower(area.Tag)
-	}
-	if filepath.IsAbs(bp) {
-		return bp
-	}
-	return filepath.Join(mm.dataPath, bp)
-}
-
-// GetAreaByID retrieves a message area by its ID.
-func (mm *MessageManager) GetAreaByID(id int) (*MessageArea, bool) {
-	mm.mu.RLock()
-	defer mm.mu.RUnlock()
-	area, exists := mm.areasByID[id]
-	return area, exists
-}
-
-// GetAreaByTag retrieves a message area by its tag.
-func (mm *MessageManager) GetAreaByTag(tag string) (*MessageArea, bool) {
-	mm.mu.RLock()
-	defer mm.mu.RUnlock()
-	area, exists := mm.areasByTag[tag]
-	return area, exists
-}
-
-// GetAreaByEchoTag retrieves a message area by its FTN echo tag.
-// Used when areas have a local tag-prefix (e.g. Tag="FD_LINUX", EchoTag="LINUX").
-func (mm *MessageManager) GetAreaByEchoTag(echoTag string) (*MessageArea, bool) {
-	mm.mu.RLock()
-	defer mm.mu.RUnlock()
-	area, exists := mm.areasByEchoTag[echoTag]
-	return area, exists
-}
-
-// UpdateAreaByID replaces the message area with the given ID with a copy of updated.
-// Callers must not modify areas by writing through pointers returned from GetAreaByID;
-// use this method so the update is performed under the manager's lock and avoids races.
-// Returns ErrAreaNotFound if no area has the given ID. The updated area's ID must match id.
-func (mm *MessageManager) UpdateAreaByID(id int, updated MessageArea) error {
-	if updated.ID != id {
-		return fmt.Errorf("message area ID mismatch: got %d, want %d", updated.ID, id)
-	}
-	mm.mu.Lock()
-	defer mm.mu.Unlock()
-	old, exists := mm.areasByID[id]
-	if !exists {
-		return ErrAreaNotFound
-	}
-	oldTag := old.Tag
-	oldEchoTag := old.EchoTag
-	replacement := new(MessageArea)
-	*replacement = updated
-	if oldTag != updated.Tag {
-		if existing, ok := mm.areasByTag[updated.Tag]; ok && existing.ID != id {
-			return fmt.Errorf("tag %q already in use by area %d", updated.Tag, existing.ID)
-		}
-		delete(mm.areasByTag, oldTag)
-	}
-	// Keep areasByEchoTag in sync when EchoTag changes.
-	if oldEchoTag != "" && oldEchoTag != old.Tag {
-		delete(mm.areasByEchoTag, oldEchoTag)
-	}
-	if updated.EchoTag != "" && updated.EchoTag != updated.Tag {
-		mm.areasByEchoTag[updated.EchoTag] = replacement
-	}
-	mm.areasByID[id] = replacement
-	mm.areasByTag[updated.Tag] = replacement
-	return nil
-}
-
-// AddArea inserts a new message area, auto-assigning the next available ID
-// and Position. The area's Tag must be unique. After insertion the area list
-// is persisted to disk. Returns the assigned ID.
-func (mm *MessageManager) AddArea(area MessageArea) (int, error) {
-	mm.mu.Lock()
-
-	// Check tag uniqueness.
-	if _, exists := mm.areasByTag[area.Tag]; exists {
-		mm.mu.Unlock()
-		return 0, fmt.Errorf("message area tag %q already exists", area.Tag)
-	}
-
-	// Assign next ID and position.
-	maxID := 0
-	maxPos := 0
-	for _, a := range mm.areasByID {
-		if a.ID > maxID {
-			maxID = a.ID
-		}
-		if a.Position > maxPos {
-			maxPos = a.Position
-		}
-	}
-	area.ID = maxID + 1
-	area.Position = maxPos + 1
-
-	// Default base path if empty.
-	if area.BasePath == "" {
-		area.BasePath = fmt.Sprintf("msgbases/area_%d", area.ID)
-	}
-
-	ptr := new(MessageArea)
-	*ptr = area
-	mm.areasByID[area.ID] = ptr
-	mm.areasByTag[area.Tag] = ptr
-	if area.EchoTag != "" && area.EchoTag != area.Tag {
-		mm.areasByEchoTag[area.EchoTag] = ptr
-	}
-	mm.mu.Unlock()
-
-	slog.Info("auto-created message area", "id", area.ID, "tag", area.Tag, "type", area.AreaType)
-
-	if err := mm.SaveAreas(); err != nil {
-		// Rollback in-memory state so it stays consistent with disk.
-		mm.mu.Lock()
-		delete(mm.areasByID, area.ID)
-		delete(mm.areasByTag, area.Tag)
-		if area.EchoTag != "" && area.EchoTag != area.Tag {
-			delete(mm.areasByEchoTag, area.EchoTag)
-		}
-		mm.mu.Unlock()
-		slog.Error("rolling back area after save failure", "tag", area.Tag, "error", err)
-		return 0, fmt.Errorf("save areas after add: %w", err)
-	}
-	return area.ID, nil
-}
-
-// SaveAreas persists all message areas to message_areas.json atomically.
-// The file is written to a temp file alongside the target and then renamed
-// to avoid partial-write corruption.
-func (mm *MessageManager) SaveAreas() error {
-	mm.mu.RLock()
-	list := make([]MessageArea, 0, len(mm.areasByID))
-	for _, area := range mm.areasByID {
-		if area != nil {
-			list = append(list, *area)
-		}
-	}
-	mm.mu.RUnlock()
-
-	sort.Slice(list, func(i, j int) bool {
-		if list[i].Position != list[j].Position {
-			return list[i].Position < list[j].Position
-		}
-		return list[i].ID < list[j].ID
-	})
-
-	data, err := json.MarshalIndent(list, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal message areas: %w", err)
-	}
-	data = append(data, '\n')
-
-	dir := filepath.Dir(mm.areasPath)
-	tmp, err := os.CreateTemp(dir, "message_areas_*.json.tmp")
-	if err != nil {
-		return fmt.Errorf("failed to create temp file for areas: %w", err)
-	}
-	tmpName := tmp.Name()
-
-	if _, err = tmp.Write(data); err != nil {
-		_ = tmp.Close()
-		_ = os.Remove(tmpName)
-		return fmt.Errorf("failed to write temp areas file: %w", err)
-	}
-	if err = tmp.Sync(); err != nil {
-		_ = tmp.Close()
-		_ = os.Remove(tmpName)
-		return fmt.Errorf("failed to sync temp areas file: %w", err)
-	}
-	if err = tmp.Close(); err != nil {
-		_ = os.Remove(tmpName)
-		return fmt.Errorf("failed to close temp areas file: %w", err)
-	}
-	if err = os.Rename(tmpName, mm.areasPath); err != nil {
-		_ = os.Remove(tmpName)
-		return fmt.Errorf("failed to rename temp areas file: %w", err)
-	}
-	return nil
-}
-
-// ListAreas returns all loaded areas sorted by Position.
-func (mm *MessageManager) ListAreas() []*MessageArea {
-	mm.mu.RLock()
-	defer mm.mu.RUnlock()
-
-	list := make([]*MessageArea, 0, len(mm.areasByID))
-	for _, area := range mm.areasByID {
-		list = append(list, area)
-	}
-	sort.Slice(list, func(i, j int) bool {
-		if list[i].Position != list[j].Position {
-			return list[i].Position < list[j].Position
-		}
-		return list[i].ID < list[j].ID
-	})
-	return list
-}
-
-// MoveAreaPosition moves the area with the given ID to newPosition,
-// shifting other areas as needed and renumbering all positions sequentially.
-func (mm *MessageManager) MoveAreaPosition(areaID int, newPosition int) error {
-	mm.mu.Lock()
-	defer mm.mu.Unlock()
-
-	area, exists := mm.areasByID[areaID]
-	if !exists {
-		return ErrAreaNotFound
-	}
-
-	// Build sorted list of all areas by current position
-	areas := make([]*MessageArea, 0, len(mm.areasByID))
-	for _, a := range mm.areasByID {
-		areas = append(areas, a)
-	}
-	sort.Slice(areas, func(i, j int) bool {
-		return areas[i].Position < areas[j].Position
-	})
-
-	// Remove the area from its current position
-	var filtered []*MessageArea
-	for _, a := range areas {
-		if a.ID != area.ID {
-			filtered = append(filtered, a)
-		}
-	}
-
-	// Clamp newPosition to valid range (1-based)
-	if newPosition < 1 {
-		newPosition = 1
-	}
-	insertIdx := newPosition - 1
-	if insertIdx > len(filtered) {
-		insertIdx = len(filtered)
-	}
-
-	// Insert at new position
-	result := make([]*MessageArea, 0, len(filtered)+1)
-	result = append(result, filtered[:insertIdx]...)
-	result = append(result, area)
-	result = append(result, filtered[insertIdx:]...)
-
-	// Renumber sequentially
-	for i, a := range result {
-		a.Position = i + 1
-	}
-
-	return nil
-}
-
-// MoveAreaPositionInConference reorders an area within its conference.
-// newIndex is 1-based within the conference's area list. Only areas in
-// the same conference are rearranged; areas in other conferences keep
-// their relative positions.
-func (mm *MessageManager) MoveAreaPositionInConference(areaID int, newIndex int) error {
-	mm.mu.Lock()
-	defer mm.mu.Unlock()
-
-	area, exists := mm.areasByID[areaID]
-	if !exists {
-		return ErrAreaNotFound
-	}
-	confID := area.ConferenceID
-
-	// Build sorted list of ALL areas by current position
-	all := make([]*MessageArea, 0, len(mm.areasByID))
-	for _, a := range mm.areasByID {
-		all = append(all, a)
-	}
-	sort.Slice(all, func(i, j int) bool {
-		return all[i].Position < all[j].Position
-	})
-
-	// Extract areas in this conference (preserving order)
-	var confAreas []*MessageArea
-	for _, a := range all {
-		if a.ConferenceID == confID {
-			confAreas = append(confAreas, a)
-		}
-	}
-
-	// Remove the source area from the conference list
-	var filtered []*MessageArea
-	for _, a := range confAreas {
-		if a.ID != areaID {
-			filtered = append(filtered, a)
-		}
-	}
-
-	// Clamp newIndex to valid range (1-based)
-	if newIndex < 1 {
-		newIndex = 1
-	}
-	insertIdx := newIndex - 1
-	if insertIdx > len(filtered) {
-		insertIdx = len(filtered)
-	}
-
-	// Insert at new position within conference
-	reordered := make([]*MessageArea, 0, len(filtered)+1)
-	reordered = append(reordered, filtered[:insertIdx]...)
-	reordered = append(reordered, area)
-	reordered = append(reordered, filtered[insertIdx:]...)
-
-	// Reassemble: interleave the reordered conference areas back into
-	// the global list at the same slots they originally occupied.
-	result := make([]*MessageArea, 0, len(all))
-	confIdx := 0
-	for _, a := range all {
-		if a.ConferenceID == confID {
-			result = append(result, reordered[confIdx])
-			confIdx++
-		} else {
-			result = append(result, a)
-		}
-	}
-
-	// Renumber all positions sequentially
-	for i, a := range result {
-		a.Position = i + 1
-	}
-
-	return nil
-}
-
-// splitNetmailTo splits a "user@zone:net/node" string into the username and
-// FTN address parts. If the string doesn't contain a valid FTN address after
-// the '@', it returns the original string unchanged with an empty address.
-func splitNetmailTo(to string) (name, addr string) {
-	idx := strings.LastIndex(to, "@")
-	if idx <= 0 {
-		return to, ""
-	}
-	candidate := strings.TrimSpace(to[idx+1:])
-	if _, err := jam.ParseAddress(candidate); err != nil {
-		return to, "" // not a valid FTN address, keep as-is
-	}
-	return strings.TrimSpace(to[:idx]), candidate
-}
-
-// addMessage is the shared implementation behind AddMessage, AddMessageWithDate,
-// AddPrivateMessage, and AddReply.
-func (mm *MessageManager) addMessage(areaID int, from, to, subject, body, replyToMsgID string,
-	replyToNum int, dateTime time.Time, private bool) (int, error) {
-	b, area, err := mm.openBase(areaID)
-	if err != nil {
-		return 0, err
-	}
-
-	// Apply body transform (e.g. V3Net tearline/origin) for the local JAM copy.
-	// The original body is preserved for OnMessagePosted so the wire message
-	// carries tearline/origin as separate protocol fields, not inline.
-	jamBody := body
-	if !private && mm.BodyTransform != nil {
-		jamBody = mm.BodyTransform(areaID, body)
-	}
-
-	msg := jam.NewMessage()
-	msg.From = from
-	msg.To = to
-	msg.Subject = subject
-	msg.Text = jamBody
-	msg.DateTime = dateTime
-
-	if private {
-		msg.Header = &jam.MessageHeader{Attribute: jam.MsgPrivate | jam.MsgLocal}
-	}
-	// replyToNum is a 1-based JAM message number; 0 means "no parent".
-	if replyToNum > 0 {
-		msg.ReplyTo = uint32(replyToNum)
-	}
-	if replyToMsgID != "" {
-		msg.ReplyID = replyToMsgID
-	}
-
-	msgType := jam.DetermineMessageType(area.AreaType, area.EchoTag)
-
-	// For netmail, split "user@address" into separate To and DestAddr fields.
-	if msgType.IsNetmail() {
-		name, addr := splitNetmailTo(to)
-		msg.To = name
-		if addr != "" {
-			msg.DestAddr = addr
-		}
-	}
-
-	var msgNum int
-	if msgType.IsEchomail() || msgType.IsNetmail() {
-		msg.OrigAddr = area.OriginAddr
-		msgNum, err = b.WriteMessageExt(msg, msgType, area.EchoTag, mm.boardName, mm.tearlineForNetwork(area.Network))
-	} else {
-		msgNum, err = b.WriteMessage(msg)
-	}
-
-	// Close the base before firing the callback. The V3Net hook calls
-	// MarkMessageSent which re-opens the same JAM base, so having it
-	// still open here can cause nested-open/file-sharing issues.
-	cerr := b.Close()
-
-	// The write already succeeded at this point, so run the post-write hooks
-	// even if the close failed — the message is on disk and downstream
-	// consumers (thread index, V3Net delivery) must still see it. The close
-	// error is folded into the returned error afterwards.
-	if err == nil {
-		mm.invalidateThreadIndex(areaID)
-		if !private && mm.OnMessagePosted != nil {
-			mm.OnMessagePosted(area, msgNum, from, to, subject, body)
-		}
-		if cerr != nil {
-			err = fmt.Errorf("closing message base: %w", cerr)
-		}
-	}
-	return msgNum, err
-}
-
-// AddMessage creates and writes a new message to the specified area.
-// For echomail areas, it automatically handles MSGID, kludges, tearline, and origin.
-// For netmail areas, "user@zone:net/node" in the To field is automatically split
-// into the username and destination address.
-// Returns the 1-based message number assigned.
-func (mm *MessageManager) AddMessage(areaID int, from, to, subject, body, replyToMsgID string) (int, error) {
-	return mm.addMessage(areaID, from, to, subject, body, replyToMsgID, 0, time.Now(), false)
-}
-
-// AddMessageWithDate is like AddMessage but uses the provided timestamp instead
-// of time.Now(). Used by V3Net to preserve the original authored date.
-func (mm *MessageManager) AddMessageWithDate(areaID int, from, to, subject, body, replyToMsgID string, dateTime time.Time) (int, error) {
-	return mm.addMessage(areaID, from, to, subject, body, replyToMsgID, 0, dateTime, false)
-}
-
-// AddPrivateMessage creates and writes a new private (MSG_PRIVATE) message.
-// For netmail areas, "user@zone:net/node" in the To field is automatically split
-// into the username and destination address. Returns the 1-based message number.
-func (mm *MessageManager) AddPrivateMessage(areaID int, from, to, subject, body, replyToMsgID string) (int, error) {
-	return mm.addMessage(areaID, from, to, subject, body, replyToMsgID, 0, time.Now(), true)
-}
-
-// AddReply creates and writes a reply, recording the parent's message number
-// (replyToNum) as the JAM ReplyTo so local areas thread, and the parent's FTN
-// MSGID (replyToMsgID, may be "") as ReplyID so echomail links via jam.Link().
-func (mm *MessageManager) AddReply(areaID int, from, to, subject, body, replyToMsgID string, replyToNum int) (int, error) {
-	return mm.addMessage(areaID, from, to, subject, body, replyToMsgID, replyToNum, time.Now(), false)
-}
-
-// AddPrivateReply is AddReply for private mail: it records the parent pointer
-// while preserving the MSG_PRIVATE flag, so a reply to a private message stays
-// private (and therefore visible to its recipient).
-func (mm *MessageManager) AddPrivateReply(areaID int, from, to, subject, body, replyToMsgID string, replyToNum int) (int, error) {
-	return mm.addMessage(areaID, from, to, subject, body, replyToMsgID, replyToNum, time.Now(), true)
-}
-
-// GetMessage reads a single message by area ID and 1-based message number.
-func (mm *MessageManager) GetMessage(areaID, msgNum int) (*DisplayMessage, error) {
-	b, _, err := mm.openBase(areaID)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		if cerr := b.Close(); cerr != nil {
-			slog.Warn("closing JAM base", "error", cerr)
-		}
-	}()
-
-	msg, err := b.ReadMessage(msgNum)
-	if err != nil {
-		return nil, err
-	}
-
-	replyToNum := 0
-	if msg.Header != nil && msg.Header.ReplyTo > 0 {
-		replyToNum = int(msg.Header.ReplyTo)
-	}
-
-	return &DisplayMessage{
-		MsgNum:     msgNum,
-		From:       msg.From,
-		To:         msg.To,
-		Subject:    msg.Subject,
-		DateTime:   msg.DateTime,
-		Body:       stripKludgeLines(normalizeLineEndings(msg.Text)),
-		MsgID:      msg.MsgID,
-		ReplyID:    msg.ReplyID,
-		ReplyToNum: replyToNum,
-		OrigAddr:   msg.OrigAddr,
-		DestAddr:   msg.DestAddr,
-		Attributes: msg.GetAttribute(),
-		IsPrivate:  msg.IsPrivate(),
-		IsDeleted:  msg.IsDeleted(),
-		AreaID:     areaID,
-	}, nil
-}
-
-// GetMessageCountForArea returns the total message count for an area.
-func (mm *MessageManager) GetMessageCountForArea(areaID int) (int, error) {
-	b, _, err := mm.openBase(areaID)
-	if err != nil {
-		if errors.Is(err, ErrAreaNotFound) {
-			return 0, nil // Return 0 if area doesn't exist
-		}
-		return 0, err // Propagate I/O and other errors
-	}
-	defer func() {
-		if cerr := b.Close(); cerr != nil {
-			slog.Warn("closing JAM base", "error", cerr)
-		}
-	}()
-
-	return b.GetMessageCount()
-}
-
-// GetTotalMessageCount returns the total number of messages across all areas.
-func (mm *MessageManager) GetTotalMessageCount() int {
-	areas := mm.ListAreas()
-	total := 0
-	for _, area := range areas {
-		count, err := mm.GetMessageCountForArea(area.ID)
-		if err != nil {
-			continue
-		}
-		total += count
-	}
-	return total
-}
-
-// GetThreadReplyCount returns the number of other messages in the same thread.
-// Threading matches Vision-2/Pascal behavior: subject-based, ignoring "Re:" prefixes
-// and " -Re: #N-" suffixes.
-func (mm *MessageManager) GetThreadReplyCount(areaID int, msgNum int, subject string) (int, error) {
-	b, _, err := mm.openBase(areaID)
-	if err != nil {
-		if errors.Is(err, ErrAreaNotFound) {
-			return 0, nil // Return 0 if area doesn't exist
-		}
-		return 0, err // Propagate I/O and other errors
-	}
-	defer func() {
-		if cerr := b.Close(); cerr != nil {
-			slog.Warn("closing JAM base", "error", cerr)
-		}
-	}()
-
-	mm.mu.RLock()
-	idx := mm.threadIndex[areaID]
-	mm.mu.RUnlock()
-
-	total, err := b.GetMessageCount()
-	if err != nil {
-		return 0, err
-	}
-
-	modCounter := uint32(0)
-	modCounterErr := false
-	if mc, err := b.GetModCounter(); err == nil {
-		modCounter = mc
-	} else {
-		modCounterErr = true
-	}
-
-	if idx == nil || idx.total != total || modCounterErr || (modCounter != 0 && idx.modCounter != modCounter) {
-		// Acquire write lock so only one goroutine rebuilds the index;
-		// others will wait and reuse the result.
-		mm.mu.Lock()
-		// Re-check after acquiring write lock (another goroutine may have rebuilt it)
-		idx = mm.threadIndex[areaID]
-		if idx == nil || idx.total != total || modCounterErr || (modCounter != 0 && idx.modCounter != modCounter) {
-			idx = mm.buildThreadIndex(b, total, modCounter)
-			mm.threadIndex[areaID] = idx
-		}
-		mm.mu.Unlock()
-	}
-
-	key := ThreadKey(subject)
-	count := idx.counts[key]
-	if count <= 1 {
-		return 0, nil
-	}
-	return count - 1, nil
-}
-
-func (mm *MessageManager) buildThreadIndex(b *jam.Base, total int, modCounter uint32) *threadIndex {
-	counts := make(map[string]int)
-	for i := 1; i <= total; i++ {
-		hdr, err := b.ReadMessageHeader(i)
-		if err != nil {
-			slog.Warn("failed to read message header", "index", i, "error", err)
-			continue
-		}
-		if hdr.Attribute&jam.MsgDeleted != 0 {
-			continue
-		}
-		subject := subjectFromHeader(hdr)
-		key := ThreadKey(subject)
-		counts[key]++
-	}
-	return &threadIndex{
-		total:      total,
-		modCounter: modCounter,
-		counts:     counts,
-	}
-}
-
-func subjectFromHeader(hdr *jam.MessageHeader) string {
-	for _, sf := range hdr.Subfields {
-		if sf.LoID == jam.SfldSubject {
-			return string(sf.Buffer)
-		}
-	}
-	return ""
-}
-
-func (mm *MessageManager) invalidateThreadIndex(areaID int) {
-	mm.mu.Lock()
-	defer mm.mu.Unlock()
-	delete(mm.threadIndex, areaID)
-	delete(mm.msgidIndex, areaID)
-}
-
-// FindMessageByMSGID searches for a message in the given area whose MSGID
-// matches the supplied value. Returns the 1-based message number, or 0 if
-// not found.  Uses a cached index that is rebuilt only when the message
-// base has changed (same invalidation strategy as threadIndex).
-func (mm *MessageManager) FindMessageByMSGID(areaID int, msgID string) int {
-	if msgID == "" {
-		return 0
-	}
-
-	b, _, err := mm.openBase(areaID)
-	if err != nil {
-		return 0
-	}
-	defer func() {
-		if cerr := b.Close(); cerr != nil {
-			slog.Warn("closing JAM base", "error", cerr)
-		}
-	}()
-
-	total, err := b.GetMessageCount()
-	if err != nil || total == 0 {
-		return 0
-	}
-
-	modCounter := uint32(0)
-	if mc, mcErr := b.GetModCounter(); mcErr == nil {
-		modCounter = mc
-	}
-
-	// Fast path: check existing index under read lock
-	mm.mu.RLock()
-	idx := mm.msgidIndex[areaID]
-	mm.mu.RUnlock()
-
-	if idx == nil || idx.total != total || (modCounter != 0 && idx.modCounter != modCounter) {
-		mm.mu.Lock()
-		// Re-check after acquiring write lock
-		idx = mm.msgidIndex[areaID]
-		if idx == nil || idx.total != total || (modCounter != 0 && idx.modCounter != modCounter) {
-			idx = mm.buildMSGIDIndex(b, total, modCounter)
-			mm.msgidIndex[areaID] = idx
-		}
-		mm.mu.Unlock()
-	}
-
-	if n, ok := idx.msgIDs[msgID]; ok {
-		return n
-	}
-	return 0
-}
-
-// buildMSGIDIndex scans all messages and builds a MSGID -> message number map.
-func (mm *MessageManager) buildMSGIDIndex(b *jam.Base, total int, modCounter uint32) *msgidIndex {
-	ids := make(map[string]int, total)
-	for i := 1; i <= total; i++ {
-		hdr, err := b.ReadMessageHeader(i)
-		if err != nil {
-			slog.Warn("failed to read message header for MSGID index", "index", i, "error", err)
-			continue
-		}
-		if hdr.Attribute&jam.MsgDeleted != 0 {
-			continue
-		}
-		for _, sf := range hdr.Subfields {
-			if sf.LoID == jam.SfldMsgID {
-				full := string(sf.Buffer)
-				ids[full] = i
-				// FTN MSGIDs are "address serial" — some tossers store REPLY
-				// kludges without the serial suffix.  Index the address part
-				// too so prefix-based lookups succeed.
-				if idx := strings.LastIndex(full, " "); idx > 0 {
-					prefix := full[:idx]
-					if _, exists := ids[prefix]; !exists {
-						ids[prefix] = i
-					}
-				}
-				break
-			}
-		}
-	}
-	return &msgidIndex{
-		total:      total,
-		modCounter: modCounter,
-		msgIDs:     ids,
-	}
-}
-
-// GetNewMessageCount returns the number of unread messages for a user in an area.
-func (mm *MessageManager) GetNewMessageCount(areaID int, username string) (int, error) {
-	b, _, err := mm.openBase(areaID)
-	if err != nil {
-		if errors.Is(err, ErrAreaNotFound) {
-			return 0, nil // Return 0 if area doesn't exist
-		}
-		return 0, err // Propagate I/O and other errors
-	}
-	defer func() {
-		if cerr := b.Close(); cerr != nil {
-			slog.Warn("closing JAM base", "error", cerr)
-		}
-	}()
-
-	return b.GetUnreadCount(username)
-}
-
-// GetLastRead returns the last-read message number for a user in an area.
-// Returns 0 if the user has no lastread record.
-func (mm *MessageManager) GetLastRead(areaID int, username string) (int, error) {
-	b, _, err := mm.openBase(areaID)
-	if err != nil {
-		if errors.Is(err, ErrAreaNotFound) {
-			return 0, nil // Return 0 if area doesn't exist
-		}
-		return 0, err // Propagate I/O and other errors
-	}
-	defer func() {
-		if cerr := b.Close(); cerr != nil {
-			slog.Warn("closing JAM base", "error", cerr)
-		}
-	}()
-
-	lr, err := b.GetLastRead(username)
-	if err != nil {
-		if err == jam.ErrNotFound {
-			return 0, nil
-		}
-		return 0, err
-	}
-	return int(lr.LastReadMsg), nil
-}
-
-// SetLastRead updates the lastread pointer for a user in an area.
-func (mm *MessageManager) SetLastRead(areaID int, username string, msgNum int) error {
-	b, _, err := mm.openBase(areaID)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if cerr := b.Close(); cerr != nil {
-			slog.Warn("closing JAM base", "error", cerr)
-		}
-	}()
-
-	return b.MarkMessageRead(username, msgNum)
-}
-
-// MarkMessageSent sets the MSG_SENT attribute on a message header.
-// Used by V3Net to indicate a locally-posted message was transmitted to the hub.
-func (mm *MessageManager) MarkMessageSent(areaID, msgNum int) error {
-	b, _, err := mm.openBase(areaID)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if cerr := b.Close(); cerr != nil {
-			slog.Warn("closing JAM base", "error", cerr)
-		}
-	}()
-
-	hdr, err := b.ReadMessageHeader(msgNum)
-	if err != nil {
-		return fmt.Errorf("mark sent: read header: %w", err)
-	}
-
-	hdr.Attribute |= jam.MsgSent
-	return b.UpdateMessageHeader(msgNum, hdr)
-}
-
-// GetNextUnreadMessage returns the next unread message number for a user.
-// Returns 0, nil if there are no unread messages.
-func (mm *MessageManager) GetNextUnreadMessage(areaID int, username string) (int, error) {
-	b, _, err := mm.openBase(areaID)
-	if err != nil {
-		if errors.Is(err, ErrAreaNotFound) {
-			return 0, nil // Return 0 if area doesn't exist
-		}
-		return 0, err // Propagate I/O and other errors
-	}
-	defer func() {
-		if cerr := b.Close(); cerr != nil {
-			slog.Warn("closing JAM base", "error", cerr)
-		}
-	}()
-
-	next, err := b.GetNextUnreadMessage(username)
-	if err != nil {
-		if err == jam.ErrNotFound {
-			return 0, nil
-		}
-		return 0, err
-	}
-	return next, nil
-}
-
-// DeleteMessage marks a message as deleted in the JAM base.
-// The message is flagged MsgDeleted; call PackAndLinkArea afterward to
-// physically remove and re-link, or run v3mail pack + link later.
-func (mm *MessageManager) DeleteMessage(areaID, msgNum int) error {
-	b, _, err := mm.openBase(areaID)
-	if err != nil {
-		return fmt.Errorf("open base for area %d: %w", areaID, err)
-	}
-	defer func() {
-		if cerr := b.Close(); cerr != nil {
-			slog.Warn("closing JAM base", "error", cerr)
-		}
-	}()
-	if err := b.DeleteMessage(msgNum); err != nil {
-		return fmt.Errorf("delete message %d in area %d: %w", msgNum, areaID, err)
-	}
-	// Invalidate caches so subsequent reads reflect the deletion
-	mm.invalidateThreadIndex(areaID)
-	return nil
-}
-
-// PackAndLinkArea packs the JAM base for the given area (removing deleted
-// messages and renumbering) then rebuilds reply threading chains. Caches
-// are invalidated afterward.
-func (mm *MessageManager) PackAndLinkArea(areaID int) error {
-	b, _, err := mm.openBase(areaID)
-	if err != nil {
-		return fmt.Errorf("open base for area %d: %w", areaID, err)
-	}
-	defer func() {
-		if cerr := b.Close(); cerr != nil {
-			slog.Warn("closing JAM base", "error", cerr)
-		}
-	}()
-	if _, err := b.Pack(); err != nil {
-		return fmt.Errorf("pack area %d: %w", areaID, err)
-	}
-	if _, err := b.Link(); err != nil {
-		return fmt.Errorf("link area %d: %w", areaID, err)
-	}
-	mm.invalidateThreadIndex(areaID)
 	return nil
 }
